+++ app/src/main/java/io/github/brainboxemb/eventtiming/app/TimingApplicationLifecycle.java
@@ -1,5 +1,7 @@
 package io.github.brainboxemb.eventtiming.app;
 
+import io.github.brainboxemb.eventtiming.application.BuildIdentity;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;